Fixed Collisions
I was not actually updating the rect positions of the tanks - just drawing them in the positions rather than moving the tank to the position I wanted and drawing its location there. But by passing in self.rect rather than self.rect.x and self.rect.y and by definging the self.rect.y of the tanks I was able to overcome this error - with LCDR Elsberry's assistance.