Fully Functioning Game Over + Play Again button
When the number of lives hits zero the end game loop is called which paints the screen red and displays game over. It also displays the play again button and checks for events so that if the button is pressed the main game loop is called again after all settings have been restored to what they were at the start of the game.