Functioning Display of lives left
I initially made this more complicated than it has to be. To get this feature to work I just need to make a new class, import the class to my game file, and then make an instance of the class for every life we have left and then just draw the group to the screen. Had to tweak the positioning by trial and error until I got the jets placed on the screen where I wanted with the size of jet I wanted.