Functioning Bomb Ground Explosion
When the bomb's y reaches the goal height that gives it the effect of hitting the ground its rectangle will be set to the goal height and its image changed to an explosion. There is a counter updated in the update so that when the counter reaches a certain number the explosion will disappear, the bomb sprite will delete itself.